+++ bin/delete_sensor.py
@@ -4,7 +4,7 @@
 # Please do not change the two lines above. See PEP 8, PEP 263.
 '''Delete an object of type: sensor'''
 __author__ = 'Jim Olsen <jim.olsen@tanium.com>'
-__version__ = '2.1.4'
+__version__ = '2.1.5'
 
 import os
 import sys