+++ CHANGELOG.md
@@ -1,6 +1,24 @@
-# 2.1.5
+# 2.1.6
 
 [Development branch](https://github.com/tanium/pytan/tree/develop) on N/A
+
+# 2.1.5
+
+[Released](https://github.com/tanium/pytan/releases/tag/2.1.5) on Sep 15 2015 
+
+  * bug fix:
+    * pollers.QuestionPoller(): increase number of get_result_info retries from 2 to 10 and add one second delay between each retry in get_result_info(), add gri_retry_count to parameter docs everywhere
+    * sessions.Session(): missing statslog in setup_logging
+  * enhancement:
+    * Added support for PyTan User Config for storing Handler arguments in JSON format
+    * handler.Handler(): add pytan_user_config parameter to init, add support for overriding init parameters with values from PyTan User Config as json file
+    * constants: add PYTAN_USER_CONFIG for default location of PyTan User Config as ~/.pytan_config.json
+    * binsupport.input_prompts(): no longer prompt for missing inputs if PyTan User Config is valid and has the parameters
+    * adding --pytan_user_config to all bin/ scripts for Handler Options
+    * handler.Handler(): added write_pytan_user_config() and read_pytan_user_config()
+    * added password obfuscation/de-obfuscation, write_pytan_user_config() will now obfsucate the password, init() will now always try to de-obfuscate the password if the password starts and ends with ::
+    * add examples for new write_pytan_user_config script
+    * re-built documentation for 2.1.5
 
 # 2.1.4